+++ test/td_qx/search/store_test.exs
@@ -1,12 +1,37 @@
-defmodule TdDd.Search.StoreTest do
+defmodule TdQx.Search.StoreTest do
   use TdQx.DataCase
 
   alias TdCluster.TestHelpers.TdDdMock
   alias TdCluster.TestHelpers.TdDfMock
+  alias TdQx.QualityControls.QualityControlVersion
   alias TdQx.Scores.ScoreGroup
   alias TdQx.Search.Store
 
-  @test_template %{
+  @template %{
+    id: 1,
+    content: [
+      %{
+        "fields" => [
+          %{
+            "cardinality" => "?",
+            "default" => %{"origin" => "default", "value" => ""},
+            "label" => "control quality",
+            "name" => "control quality",
+            "type" => "string",
+            "values" => nil,
+            "widget" => "string"
+          }
+        ],
+        "name" => ""
+      }
+    ],
+    label: "quality control",
+    name: "quality control",
+    scope: "quality_control",
+    subscope: nil
+  }
+
+  @template_score %{
     id: 5,
     name: "foo",
     label: "score_group",
@@ -49,7 +74,90 @@ defmodule TdDd.Search.StoreTest do
     ]
   }
 
-  describe "Store.stream/1" do
+  describe "stream/1" do
+    test "streams over all quality control versions" do
+      quality_control = insert(:quality_control)
+
+      TdDfMock.list_templates_by_scope(&Mox.expect/4, "quality_control", {:ok, [@template]})
+      TdDdMock.log_start_stream(&Mox.expect/4, 2, :ok)
+      TdDdMock.log_progress(&Mox.expect/4, 1, :ok)
+      TdDdMock.log_progress(&Mox.expect/4, 1, :ok)
+
+      versioned =
+        insert(:quality_control_version,
+          quality_control: quality_control,
+          status: "versioned",
+          version: 1
+        )
+
+      score_versioned = insert(:score, quality_control_version: versioned)
+      insert(:score_event, type: "FAILED", score: score_versioned)
+
+      published =
+        insert(:quality_control_version,
+          quality_control: quality_control,
+          df_type: "quality control",
+          status: "published",
+          version: 2
+        )
+
+      score_published = insert(:score, quality_control_version: published)
+      insert(:score_event, type: "SUCCEEDED", score: score_published)
+
+      {:ok, to_index} =
+        Repo.transaction(fn ->
+          QualityControlVersion
+          |> Store.stream()
+          |> Enum.to_list()
+        end)
+
+      assert Enum.count(to_index) == 2
+      assert versioned = Enum.find(to_index, &(&1.id == versioned.id))
+      assert versioned.latest_score.id == score_versioned.id
+      assert versioned.latest_score.status == "FAILED"
+      refute versioned.latest
+      assert published = Enum.find(to_index, &(&1.id == published.id))
+      assert published.latest
+      assert published.latest_score.id == score_published.id
+      assert published.latest_score.status == "SUCCEEDED"
+    end
+  end
+
+  describe "stream/2" do
+    test "streams over quality control versions provided a list of ids" do
+      TdDfMock.list_templates_by_scope(&Mox.expect/4, "quality_control", {:ok, [@template]})
+      TdDdMock.log_start_stream(&Mox.expect/4, 1, :ok)
+      TdDdMock.log_progress(&Mox.expect/4, 1, :ok)
+
+      _version1 =
+        insert(:quality_control_version,
+          quality_control: build(:quality_control),
+          df_type: "quality control",
+          status: "published",
+          version: 2
+        )
+
+      version2 =
+        insert(:quality_control_version,
+          quality_control: build(:quality_control),
+          df_type: "quality control",
+          status: "published",
+          version: 2
+        )
+
+      {:ok, [to_index]} =
+        Repo.transaction(fn ->
+          QualityControlVersion
+          |> Store.stream(quality_control_ids: [version2.quality_control_id])
+          |> Enum.to_list()
+        end)
+
+      assert to_index.id == version2.id
+      assert to_index.template == @template
+      assert to_index.status == "published"
+      assert to_index.latest
+    end
+
     setup do
       Application.put_env(Store, :chunk_size, 10)
       :ok
@@ -87,7 +195,7 @@ defmodule TdDd.Search.StoreTest do
       TdDfMock.list_templates_by_scope(
         &Mox.expect/4,
         "qxe",
-        {:ok, [@test_template]}
+        {:ok, [@template_score]}
       )
 
       TdDdMock.log_start_stream(
@@ -103,7 +211,7 @@ defmodule TdDd.Search.StoreTest do
         3
       )
 
-      template = @test_template
+      template = @template_score
 
       assert [
                %{
@@ -160,7 +268,7 @@ defmodule TdDd.Search.StoreTest do
       TdDfMock.list_templates_by_scope(
         &Mox.expect/4,
         "qxe",
-        {:ok, [@test_template]}
+        {:ok, [@template_score]}
       )
 
       TdDdMock.log_start_stream(
@@ -176,7 +284,7 @@ defmodule TdDd.Search.StoreTest do
         3
       )
 
-      template = @test_template
+      template = @template_score
 
       assert [
                %{
@@ -192,7 +300,7 @@ defmodule TdDd.Search.StoreTest do
              ] =
                Store.transaction(fn ->
                  ScoreGroup
-                 |> Store.stream([score_group_id])
+                 |> Store.stream(id: score_group_id)
                  |> Enum.to_list()
                end)
     end